// practice_typescript/Ilari/src/types.ts
export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Windy = 'windy',
  Stromy = 'stromy'
}

export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor',
}

export interface DiaryEntry {
    id: number;
    date: string;
    weather: Weather;
    visibility: Visibility;
    comment?: string;
  }

  export type NonSensitiveDiaryEntry = Omit<DiaryEntry, 'comment'>; 
  export type NewDiaryEntry = Omit<DiaryEntry, 'id'>; 